requirements.txt
fastapi
python-dotenv
uvicorn
motor
redis
pydantic
pymongo